WIP: Re-shuffle the auction results page
